requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
distlib==0.3.3
Django==3.2.9
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap4==3.0.1
django-datatable==0.3.1
django-debug-toolbar==3.2.1
django-encrypted-model-fields==0.6.0
django-environ==0.8.1
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
filelock==3.4.0
importlib-metadata==4.10.0
Markdown==3.3.6
mysqlclient==2.1.0
numpy==1.22.0
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
protobuf==3.17.3
pycodestyle==2.8.0
pycparser==2.21
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.0.0
toml==0.10.2
tqdm==4.62.3
virtualenv==20.10.0
virtualenv-clone==0.5.7
zipp==3.6.0